Day 20, part 2 incomplete
First attempt - always try part 2 using the same brute force as part 1
if you can. In this case, does not work and there is basically no
achievable optimisation - it has to be done mathematically, or I could
leave this running for a century and get my start post WWIV when I rule
the wasteland which is all that remains of civilisation.

+++ 2023/Day20.h
@@ -47,7 +47,7 @@ private:
 	const std::string testFile2 = "inputs/day20_part2_test.txt";
 	const std::string puzzleFile = "inputs/day20.txt";
 	uint64_t calculatePuzzle1(std::vector<std::string> input);
-	int calculatePuzzle2(std::vector<std::string> input);
+	uint64_t calculatePuzzle2(std::vector<std::string> input);
 public:
 	void puzzle1();
 	void puzzle2();
